make 14 homework,fixes refactoring

+++ src/blogs/blogs.service.ts
@@ -1,12 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { BlogRepository } from './blog.repository';
 import { BlogQueryRepository } from './blogQuery.repository';
-import {
-  BlogDbType,
-  BlogInputModel,
-  BlogViewType,
-  CreateBlogDto,
-} from './blogs.types';
+import { BlogDbType, BlogViewType, CreateBlogDto } from './blogs.types';
 import { ObjectId } from 'mongodb';
 import { Helpers } from '../helpers/helpers';
 
@@ -18,16 +13,15 @@ export class BlogService {
     public helpers: Helpers,
   ) {}
 
-  async deleteBlog(blogId: string): Promise<boolean> {
+  async deleteBlog(blogId: string): Promise<boolean | null> {
+    const isExistBlog = await this.blogQueryRepository.findBlogById(blogId);
+    if (!isExistBlog) return null;
+
     return await this.blogRepository.deleteBlog(blogId);
   }
 
   async createBlog(blogData: CreateBlogDto): Promise<BlogViewType> {
-    const khe = await this.helpers.validateOrRejectModel(
-      blogData,
-      CreateBlogDto,
-    );
-    console.log(khe);
+    await this.helpers.validateOrRejectModel(blogData, CreateBlogDto);
     const newBlog: BlogDbType = {
       _id: new ObjectId(),
       name: blogData.name,
@@ -41,8 +35,11 @@ export class BlogService {
 
   async updateBlog(
     blogId: string,
-    blogUpdateData: BlogInputModel,
-  ): Promise<boolean> {
+    blogUpdateData: CreateBlogDto,
+  ): Promise<boolean | null> {
+    await this.helpers.validateOrRejectModel(blogUpdateData, CreateBlogDto);
+    const isExistBlog = await this.blogQueryRepository.findBlogById(blogId);
+    if (!isExistBlog) return null;
     return await this.blogRepository.updateBlog(blogId, blogUpdateData);
   }
 }
